test(web): assert the export subject filter against the real helper
The test lived under src/routes, which main now bans by eslint because the
TanStack route generator scans that directory and reads a dot as a path
separator. It also only exercised tanstack's row model and `Set.has`, so it
would still have passed had the per-cell extraction been reintroduced.

Move `getListedSubjectIds` to src/utils/table.ts and assert it directly, and
throw when the table capture fails rather than asserting on `table!`.
